Muutin välituloksien tarkkuuden double:ksi kokonaistarkkuuden parantamiseksi. Lisäksi loppulaskussa varmistetaan, ettei neliöjuuresta tule negatiivinen pyöristysvirheiden takia. Tällöin palautetaan keskihajontana luku nolla.

+++ include/ModStandardDeviationCalculator.h
@@ -49,9 +49,9 @@ namespace WeatherAnalysis
 	const int itsModulo;
 
 	long itsCounter;
-	float itsSum;
-	float itsSquaredSum;
-	float itsPreviousDirection;
+	double itsSum;
+	double itsSquaredSum;
+	double itsPreviousDirection;
 
   }; // class ModStandardDeviationCalculator
